refactor: cleaned up connection.js separated events callbacks into functions
